Class py::oby  moved to a separate file
