+++ c/frame/__getitem__.cc
@@ -56,11 +56,11 @@
 // each subframe's RowIndex).
 //
 //------------------------------------------------------------------------------
-#include "expr/by_node.h"
+#include "expr/eval_context.h"
 #include "expr/expr.h"
 #include "expr/join_node.h"
+#include "expr/py_by.h"         // py::oby
 #include "expr/sort_node.h"
-#include "expr/eval_context.h"
 #include "frame/py_frame.h"
 #include "python/_all.h"
 #include "python/string.h"
